pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rsa_square"
version = "0.1dev"
requires-python = ">=3.10,<3.13"
dependencies = [
  "torch==2.6.0",
  "torchvision==0.21.0",
  "transformers==4.46.3",
  "accelerate==1.0.1",
  "sentencepiece==0.2.0",
  "sentence-transformers==3.4.1",
  "pandas==2.2.3",
  "numpy==2.2.3",
  "scipy==1.15.1",
  "scikit-learn==1.6.1",
  "matplotlib==3.9.2",
  "hydra-core==1.3.2",
  "omegaconf==2.3.0",
  "openpyxl==3.1.5",
  "tabulate==0.9.0",
  "requests==2.32.3",
  "tqdm==4.66.5",
  "PyYAML==6.0.2",
  "jsonlines==4.0.0",
  "regex==2024.9.11",
  "pillow==11.0.0",
  "protobuf==5.28.3"
]

[tool.setuptools]
packages = ["src"]

[project.optional-dependencies]
dev = [
  "ipykernel",
  "ruff",
  "pytest",
]
